Soul Gems, tentacle grab fixes

+++ Assets/_Project/Scripts/Characters/States/CharacterDeathState.cs
@@ -5,16 +5,28 @@ using Prime31.StateKit;
 
 public class CharacterDeathState : SKState<Character> {
 
+	float deathTime = 2f;
+	float startTime;
+
 	public override void begin ()
 	{
 		LevelManager.AddKill ();
 
-		_context.Destroy();
+		if (_context.soulGemPrefab != null)
+		{
+			GameObject go = GameObject.Instantiate(_context.soulGemPrefab);
+			go.transform.position = _context.transform.position;
+		}
+
+		_context.anim.SetTrigger("Death");
+
+		startTime = Time.time;
 	}
 
 	public override void update (float deltaTime)
 	{
-
+		if (Time.time > startTime + deathTime)
+			_context.Destroy();
 	}
 
 	public override void end ()
